backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
openpyxl==3.1.2
pandas==2.1.3
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
gitpython==3.1.40
openai==1.3.5
langchain==0.0.350
aiofiles==23.2.1
